Added Data.IOArray for mutable array primitives
The interface provides only 'unsafe' reading and writing. This gives
primtive access without bounds checking. Safe interfaces should be built
on top of this, but it's up to the implementer whether bounds checking
is at compile time (e.g. via a proof of bounds safety with the index
mapping to an integer location) or at runtime (e.g. returning a Maybe)