Remove type synonym that points Float to Double.
Several releases ago we decided to deprecate the use of `Float` in
favour of `Double` to describe 64bit floating point numbers. For the
rationale please see our FAQs [1].

Given that the decision was made several releases ago I think it is
time that we remove the type alias that point's `Float` to `Double`.

There has been enough time for people ot move on.

[1] http://idris.readthedocs.io/en/latest/faq/faq.html#why-does-idris-use-double-instead-of-float64