Set environment with Idris flags while running `make`
When building a package with foreign code and a makefile, the makefile
itself must find the compilation flags that Idris already knows
about. Moreover, the `idris` executable found by the makefile may not
be the same used to build the package. This change passes the build
flags to the makefile though environment variables.