Fix binding rules in patterns
Names should only be implicitly bound in patterns if they begin with a
lower case letter or _ (as in types), and we've always documented this
but never actually enforced it! This has led to some mysterious and
undetected bugs in Idris programs.

This is likely to break some code (sorry) given that I've had to update
a few tests to follow the new rules. However, the code it breaks is code
that shouldn't have worked in the first place!