Set C language standard to C99 in runtime and codegen (#4308)
* Fix tests for compilation under Windows

C preprocessor tests for detecting Windows were inconsistent across
the code base, and in some cases wrong. This change makes the
detection uniform and correct at all places.

* Write `usleep` in terms of `nanosleep`

`usleep` is deprecated and a compiler warning is issued if it's used
with a POSIX compiler flag set. As an added bonus, it's range has been
increased.

* Set C language standard to C99 in runtime and codegen

Currently the exact C language used by the runtime and the C code
generator is not set and therefore different compilers may use
different standards depending on their default behaviour. This may
lead to inconsistencies in builds, such a sucessful build in one
machine failing in another.