Reverse improvements (#4310)
* Expose helper for reverse.

Not only is it a useful function all its own, the lack of exposure prevented
any proofs about the behavior of reverse outside of Prelude.List.

* Fundamental properties for reverse and reverseOnto.

These are less proerties, and more specification.  These properties should be
sufficient for any proofs about the behavior of the functions, and any
implementations violating these properties would not generally be considered
correct.

* Some additional properties about reverse.

reverseAppend and reverseReverseId are proven in terms of the reverse
specification proeprties and should be robust against future implemetation
changes, or migration to a module separate from reverse.

reverseSingletonId is a trivial proof, but will likely need updating if it can
no longer "see" the reverse implemenation.