Initial infrastructure for IdrisDoc annotations.
This commit puts in place the infrastructure for IdrisDoc to support
documentation metadata. Analogous to '@' annotations used in Doxygen
and Javadoc. We do this by wrapping around `DocString` a documentation data structure that contains the relevant metadata to be capture for each core Idris constructuct.

Current constructs that accept metadata are:

1. Type constructors.
1. Data Definitions & value constructors.
1. Modules
1. Record constructors & value constructors and fields
1. Interface & implementations

The capture metadata varies between each construct but are grouped amongst:

1. Modules
1. arguments
1. things that construct other things

See Idris.Documentation for more details.

We don't do anything with the infrastructure yet, this comes later.
We still need to:

1. update the parser to correctly parse and construct different metadata groups.
1. update the renderer to display the metadata appropriately.
1. see what needs updating for IdrisDoc
1. update the IDE protocol where appropriate.

With this commit the average idris user should not notice any difference.