Reduce work allNamesIn has to do
This is a major performance hit when functions get big (especially in do
blocks) and a lot of the work it's doing is not actually necessary since
the primary goal is to find shadowed names. So, only search alternatives
if they are genuine alternatives rather than there for name
disambiguation.

Also reduce the reliance on uniplate a little bit, since this causes
some (smaller) performance problems too.