Use a newer cabal with 7.10
The 7.10 build fails when the cabal executable crashes in the configure step.

It works with a new cabal, so let's use that. Any 7.10 user can install a newer
cabal-install if it would turn out to be more than a Travis environment issue.

Update explanation for why 8.4 doesn't work with new roadblock (sigh).