Tail calls via trampolines in C back end
This means we can turn off the clang optimiser and still have working
code generated. It does mean that the code generated is a bit slower
(there's a small overhead for the trampolines) but it compiles much
faster, which is a big deal for developing larger systems!

We need a way to turn the clang optimiser back on - probably just a
compiler flag.