Why nub a list when you can use a set.
Highlighting information was contained in an association list. Not
only were there duplicates in the list, nubbing the list resulted in
performance losses on proper sized code bases.

We migrate the lists to sets and use `FC'` to ensure no false
positives are returned when comparing to elements in the list. `FC'`
is `FC` but with equality.

Anecdotal tests show good performance improvements wrt to code highlighting.