newBuffer should return a managed pointer
Otherwise, we're just guessing on the size, and that way lies segfaults.
It turns out the size in Buffer.idr might have been correct once, but
isn't now, and so unless we're lucky registering the pointer will copy
memory we don't have access to. If the RTS does the registering of the
pointer, at least it knows the size.

(It's likely this is doable without the malloc, using iallocate
directly, but this required fewer changes.)