Wrap GMP operations in a requireAlloc
Since GMP uses the idris allocator, and when it uses it is out of our
control, we need to be sure it doesn't GC in the middle and move things
it's working with. I have merely guessed at an upper limit for what
it'll need. This is a horrible hack. Sorry. If you're working with
numbers that don't fit in 64k you might have problems...