Store hash of the IBC file's interface
This saves rechecking a file if none of its imports have the same
interfaces and there are otherwise no changes.

The hash function is probably quite bad, so if you know how to do a
better one please consider contributing it! Also, this makes an already
terrible piece of code in Parser.hs even worse (sorry) and it's crying
out for a refactoring if anyone has any enthusiasm for it (as is the
whole system of chasing dependencies and loading modules).

This is mostly to speed up development of Blodwen, which was really
starting to get painful when making small changes to core modules.