Initial transfer of {AS} grammar to Sphinx Documtation.
When {AS} originally moved Idris' parser from parsec to Trifecta, a grammar describing the surface syntax was contributed as source code comments.
This is good for developers, but not necessarily helpful for those not wanting to have a reference.
Given that we have moved documentation from LaTeX to sphix-doc we can leverage the `productionlist` environment to provide a clickable grammar.

I have split the grammar into sections, but over time this should be pruned, sorted, and formatted properly.
Further I think a good contribution would be to merge this file with the proceeding section to provide examples alongside.
It might be work spinning out these sections as a top level document in its own right, with the 'Language Reference' being placed as a 'Idris Tooling Guide'.