Add Clock/clockTime in System
This is for getting the time at a finer granularity than seconds. I
want this for getting approximate wall clock time measurements for run
time of fragments of programs, so the interface is very basic, but I am
open to better, more sophisticated APIs!

I expect this won't work on Windows, so it's #ifdefed out and always
returns 0. If you know how to fix this, please do!