Show visibility in doc comments.
Currently the result of `:doc` displays totality information for functions.
However there are two scenarios where having the accessiblity/visibility information would be beneficial:

1. Type-level computations failing to reduce.

A computed value in your type might fail to reduce in the definition of the computing function is not public c.f. exportation of functions computing type synonyms.
A good numpty check would be to add this information to the doc output, much as we do for totality status.

2. Working with 'Exported' constructors.

Much like exported functions, some data types might be exported rather than public exported.
One has to assert the functions visibility by checking the source code, this is cumbersome.

This commit adds visibility status to the documentation output.
Such that if totality information is to be displayed we output the totality and accessibility information, otherwise we just output the accessibility information.