Add an alternative stack configuration file to deal with upstream GHC issues.
There is an upstream issue with GHC on some Ubuntu and Fedora machines.
The issue is that for GHC versions greater than 8.4.X linking to libFFI is broken.
See the following GHC issue page for more information:

  <https://gitlab.haskell.org/ghc/ghc/issues/15397>

This means that an Idris built with `libFFI` support will fail at runtime as the executable would not be able to find the correct `libffi` shared object.
This will cause the build to fail.

This commit adds an alternative stack configuration file to help bypass the issues for those affected by the issue.